+++ app/src/components/Map/VesselInfoPanel.jsx
@@ -5,14 +5,14 @@ import { getCountry } from 'iso-3166-1-alpha-2';
 import MediaQuery from 'react-responsive';
 import ExpandButton from 'components/Shared/ExpandButton';
 
-import vesselPanelStyles from 'styles/components/vessel-info-panel.scss';
+import infoPanelStyles from 'styles/components/info-panel.scss';
 import buttonCloseStyles from 'styles/components/button-close.scss';
 import iconStyles from 'styles/icons.scss';
 
 import CloseIcon from '-!babel-loader!svg-react-loader!assets/icons/close.svg?name=Icon';
 import PinIcon from '-!babel-loader!svg-react-loader!assets/icons/pin.svg?name=PinIcon';
 
-import { VESSEL_INFO_STATUS } from 'constants';
+import { INFO_STATUS } from 'constants';
 
 class VesselInfoPanel extends Component {
 
@@ -34,21 +34,21 @@ class VesselInfoPanel extends Component {
     const vesselInfo = this.props.currentlyShownVessel;
     const status = this.props.infoPanelStatus;
 
-    if (status !== VESSEL_INFO_STATUS.LOADING && vesselInfo === null) {
+    if (status !== INFO_STATUS.LOADING && vesselInfo === null) {
       return null;
     }
 
     let vesselInfoContents = null;
 
-    if (status === VESSEL_INFO_STATUS.LOADING) {
+    if (status === INFO_STATUS.LOADING) {
       vesselInfoContents = (
-        <div className={vesselPanelStyles.vesselMetadata} >
+        <div className={infoPanelStyles.metadata} >
           <div>Loading vessel information...</div>
         </div>
       );
     } else if (this.props.userPermissions !== null && this.props.userPermissions.indexOf('seeVesselBasicInfo') === -1) {
       return null;
-    } else if (status === VESSEL_INFO_STATUS.LOADED && vesselInfo) {
+    } else if (status === INFO_STATUS.LOADED && vesselInfo) {
       const layerFields = this.props.layerFieldsHeaders;
 
       const canSeeVesselDetails = (this.props.userPermissions !== null && this.props.userPermissions.indexOf('info') !== -1);
@@ -56,19 +56,19 @@ class VesselInfoPanel extends Component {
       const renderedFieldList = this.generateVesselDetails(layerFields, canSeeVesselDetails, vesselInfo);
 
       vesselInfoContents = (
-        <div className={vesselPanelStyles.vesselMetadata} >
+        <div className={infoPanelStyles.metadata} >
           {renderedFieldList}
           {((this.props.userPermissions !== null && this.props.userPermissions.indexOf('pin-vessel') !== -1) || vesselInfo.pinned) &&
             <PinIcon
               className={classnames(iconStyles.icon, iconStyles.pinIcon,
-                vesselPanelStyles.pinIcon, { [`${vesselPanelStyles._pinned}`]: vesselInfo.pinned })}
+                infoPanelStyles.pinIcon, { [`${infoPanelStyles._pinned}`]: vesselInfo.pinned })}
               onClick={() => {
                 this.props.onTogglePin(vesselInfo.seriesgroup);
               }}
             />
           }
           {canSeeVesselDetails && vesselInfo.mmsi && <a
-            className={vesselPanelStyles.externalLink}
+            className={infoPanelStyles.externalLink}
             target="_blank"
             rel="noopener noreferrer"
             href={`http://www.marinetraffic.com/en/ais/details/ships/mmsi:${vesselInfo.mmsi}`}
@@ -76,7 +76,7 @@ class VesselInfoPanel extends Component {
           </a>
           }
           {!canSeeVesselDetails && <a
-            className={vesselPanelStyles.externalLink}
+            className={infoPanelStyles.externalLink}
             onClick={this.props.login}
           >Click here to login and see more details</a>
           }
@@ -86,10 +86,10 @@ class VesselInfoPanel extends Component {
 
     return (
       <div
-        className={classnames(vesselPanelStyles.vesselInfoPanel,
-          { [`${vesselPanelStyles._expanded}`]: this.state.isExpanded })}
+        className={classnames(infoPanelStyles.infoPanel,
+          { [`${infoPanelStyles._expanded}`]: this.state.isExpanded })}
       >
-        <div className={vesselPanelStyles.buttonsContainer} >
+        <div className={infoPanelStyles.buttonsContainer} >
           <MediaQuery maxWidth={789} >
             <ExpandButton
               onExpand={() => this.onExpand()}
@@ -99,7 +99,7 @@ class VesselInfoPanel extends Component {
 
           <button
             onClick={() => this.props.hide()}
-            className={classnames(buttonCloseStyles.buttonClose, vesselPanelStyles.closeBtn)}
+            className={classnames(buttonCloseStyles.buttonClose, infoPanelStyles.closeBtn)}
           >
             <CloseIcon className={buttonCloseStyles.cross} />
           </button>
@@ -121,12 +121,12 @@ class VesselInfoPanel extends Component {
           if (!vesselInfo[field.id]) {
             break;
           }
-          renderedFieldList.push(<div key={field.id} className={vesselPanelStyles.rowInfo} >
-            <span className={vesselPanelStyles.key} >{field.display}</span>
-            <ul className={vesselPanelStyles.linkList} >
-              <li className={vesselPanelStyles.linkListItem} >
+          renderedFieldList.push(<div key={field.id} className={infoPanelStyles.rowInfo} >
+            <span className={infoPanelStyles.key} >{field.display}</span>
+            <ul className={infoPanelStyles.linkList} >
+              <li className={infoPanelStyles.linkListItem} >
                 <a
-                  className={vesselPanelStyles.externalLink}
+                  className={infoPanelStyles.externalLink}
                   href={`${field.prefix}${vesselInfo[field.id]}`}
                   target="_blank"
                   rel="noopener noreferrer"
@@ -140,9 +140,9 @@ class VesselInfoPanel extends Component {
         case 'objectArrayMultiLink':
           linkList = [];
           vesselInfo[field.id].forEach((registry) => {
-            linkList.push(<li key={registry.rfmo} className={vesselPanelStyles.linkListItem} >
+            linkList.push(<li key={registry.rfmo} className={infoPanelStyles.linkListItem} >
               <a
-                className={vesselPanelStyles.externalLink}
+                className={infoPanelStyles.externalLink}
                 href={`${registry.url}`}
                 target="_blank"
                 rel="noopener noreferrer"
@@ -151,23 +151,23 @@ class VesselInfoPanel extends Component {
               </a>
             </li>);
           });
-          renderedFieldList.push(<div key={field.id} className={vesselPanelStyles.rowInfo} >
-            <span className={vesselPanelStyles.key} >{field.display}</span>
-            <ul className={vesselPanelStyles.linkList} >
+          renderedFieldList.push(<div key={field.id} className={infoPanelStyles.rowInfo} >
+            <span className={infoPanelStyles.key} >{field.display}</span>
+            <ul className={infoPanelStyles.linkList} >
               {linkList}
             </ul>
           </div>);
           break;
         case 'flag':
-          renderedFieldList.push(<div key={field.id} className={vesselPanelStyles.rowInfo} >
-            <span className={vesselPanelStyles.key} >{field.display}</span>
-            <span className={vesselPanelStyles.value} >{getCountry(vesselInfo[field.id]) || '---'}</span>
+          renderedFieldList.push(<div key={field.id} className={infoPanelStyles.rowInfo} >
+            <span className={infoPanelStyles.key} >{field.display}</span>
+            <span className={infoPanelStyles.value} >{getCountry(vesselInfo[field.id]) || '---'}</span>
           </div>);
           break;
         default:
-          renderedFieldList.push(<div key={field.id} className={vesselPanelStyles.rowInfo} >
-            <span className={vesselPanelStyles.key} >{field.display}</span>
-            <span className={vesselPanelStyles.value} >{vesselInfo[field.id] || '---'}</span>
+          renderedFieldList.push(<div key={field.id} className={infoPanelStyles.rowInfo} >
+            <span className={infoPanelStyles.key} >{field.display}</span>
+            <span className={infoPanelStyles.value} >{vesselInfo[field.id] || '---'}</span>
           </div>);
       }
     });